Fix flaky test in cluster failover

+++ store/cluster_node.go
@@ -255,7 +255,9 @@ func (n *ClusterNode) SyncClusterInfo(ctx context.Context, cluster *Cluster) err
 }
 
 func (n *ClusterNode) Reset(ctx context.Context) error {
-	_ = n.GetClient().FlushAll(ctx).Err()
+	if n.IsMaster() {
+		_ = n.GetClient().FlushAll(ctx).Err()
+	}
 	return n.GetClient().ClusterResetHard(ctx).Err()
 }
